% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_p}
\alias{data_p}
\title{P value matrix of two data types}
\format{
A data matrix with 1000 rows and 2 variables:
\describe{
\item{p.rnaseq}{P values of data type 1 for all genes}
\item{p.cnv}{P values of data type 2 for all genes}
}
}
\usage{
data(data_p)
}
\description{
A dataset with summary statistics p values of 1000 genes for RNAseq and CNV data
}
\keyword{datasets}
