% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FDR_control_adaptive.R
\name{FDR_control_adaptive_imix}
\alias{FDR_control_adaptive_imix}
\title{The adaptive procedure for global FDR control for IMIX output}
\usage{
FDR_control_adaptive_imix(
  imix_output,
  model = c("IMIX_ind", "IMIX_cor_twostep", "IMIX_cor_restrict", "IMIX_cor"),
  alpha
)
}
\arguments{
\item{imix_output}{The result output from IMIX() function, result controled at alpha level only for one component each time.}

\item{model}{The target model among "IMIX_ind","IMIX_cor_twostep","IMIX_cor_restrict", and "IMIX_cor". Default is IMIX_ind.}

\item{alpha}{Prespecified FDR control level.}
}
\value{
The estimated mFDR for the target component and classify the genes in each component after FDR control at alpha level.
}
\description{
The adaptive procedure for global FDR control based on the output from IMIX models, this can be directly performed by IMIX function, however, if you are interested in other alpha levels, this function would be useful to avoid rerun the IMIX().
}
