% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMIX.R
\name{IMIX}
\alias{IMIX}
\title{IMIX}
\usage{
IMIX(
  data_input,
  data_type = c("p", "z"),
  mu_ini = NULL,
  cov_ini = NULL,
  p_ini = NULL,
  tol = 1e-06,
  maxiter = 1000,
  seed = 10,
  ini.ind = TRUE,
  model = c("all", "IMIX_ind", "IMIX_cor_twostep", "IMIX_cor_restrict", "IMIX_cor"),
  model_selection_method = c("AIC", "BIC"),
  alpha = 0.2,
  verbose = FALSE
)
}
\arguments{
\item{data_input}{An n x d data frame or matrix of the summary statistics z score or p value, n is the nubmer of genes, d is the number of data types. Each row is a gene, each column is a data type.}

\item{data_type}{Whether the input data is the p values or z scores, default is p value}

\item{mu_ini}{A list of initial values for the mean vectors for each component. If there are three data types and 8 components, then the initial is a list of 8 mean vectors, each vector is of length 3.}

\item{cov_ini}{A list of initial values for the covariance matrices. If there are three data types and 8 components, then the initial is a list of 8 covariance matrices, each matix is 3*3.}

\item{p_ini}{Initial value for the proportion of the distribution in the Gaussian mixture model. A vector of length 2^d, d is the number of data types.}

\item{tol}{The convergence criterion. Convergence is declared when the change in the observed data log-likelihood increases by less than epsilon.}

\item{maxiter}{The maximum number of iteration, default is 1000}

\item{seed}{Set.seed, default is 10}

\item{ini.ind}{Use the parameters estimated from IMIX-ind for initial values of other IMIX models, default is TRUE}

\item{model}{Which model to use to compute the data, default is all}

\item{model_selection_method}{Model selection information criteria, based on AIC or BIC, default is AIC}

\item{alpha}{Prespecified nominal level for global FDR control, default is 0.2}

\item{verbose}{Whether to print the full log-likelihood for each iteration, default is FALSE}
}
\value{
The results of IMIX
}
\description{
Fitting a multivariate mixture model framework, model selection for the best model, and adaptive procedure for FDR control. Input of summary statistics z scores of two or three data types.
}
