% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMIX_ind.R
\name{IMIX_ind}
\alias{IMIX_ind}
\title{IMIX-ind}
\usage{
IMIX_ind(
  data_input,
  data_type = c("p", "z"),
  mu,
  sigma,
  p,
  tol = 1e-06,
  maxiter = 1000,
  seed = 10,
  verbose = FALSE
)
}
\arguments{
\item{data_input}{An n x d data frame or matrix of the summary statistics z score or p value, n is the nubmer of genes, d is the number of data types. Each row is a gene, each column is a data type.}

\item{data_type}{Whether the input data is the p values or z scores, default is p value}

\item{mu}{Initial value for the mean of each component of the independent mixture model distribution. A vector of length 2*d, d is number of data types. Needs to be in a special format that corresponds to the initial value of mu, for example, if d=3, needs to be in the format of (null_1,alternative_1,null_2,alternative_2,null_3,alternative_3).}

\item{sigma}{Initial value for the standard deviation of each component of the independent mixture model distribution. A vector of length 2*d, d is number of data types. Needs to be in a special format that corresponds to the initial value of mu, for example, if d=3, needs to be in the format of (null_1,alternative_1,null_2,alternative_2,null_3,alternative_3).}

\item{p}{Initial value for the proportion of the distribution in the Gaussian mixture model}

\item{tol}{The convergence criterion. Convergence is declared when the change in the observed data log-likelihood increases by less than epsilon.}

\item{maxiter}{The maximum number of iteration, default is 1000}

\item{seed}{set.seed, default is 10}

\item{verbose}{Whether to print the full log-likelihood for each iteration, default is FALSE}
}
\value{
A list of the results of IMIX-ind
\item{posterior prob}{Posterior probability matrix of each gene for each component}
\item{Full LogLik all}{Full log-likelihood of each iteration}
\item{Full MaxLogLik final}{The final log-likelihood of the converged model}
\item{iterations}{Number of iterations run}
\item{pi}{Estimated proportion of each component, sum to 1}
\item{mu}{Estimated mean for the null and alternative of each data type: for two data types (mu10,mu11,mu20,mu21), three data types (mu10,mu11,mu20,mu21,mu30,mu31), mui0 is the null for data type i, mui1 is the alternative for data type i.}
\item{sigma}{Estimated standard deviation for the null and alternative of each data type: for two data types (sigma10,sigma11,sigma20,sigma21), three data types (sigma10,sigma11,sigma20,sigma21,sigma30,sigma31), sigmai0 is the null for data type i, sigmai1 is the alternative for data type i.}
}
\description{
Fitting an independent mixture model with restrictions on mean and variance. Input of summary statistics z scores or p values of two or three data types.
}
\references{
Wang, Ziqiao, and Peng Wei. 2020. “IMIX: A Multivariate Mixture Model Approach to Integrative Analysis of Multiple Types of Omics Data.” BioRxiv. Cold Spring Harbor Laboratory. \url{https://doi.org/10.1101/2020.06.23.167312}.
}
