% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection.R
\name{model_selection}
\alias{model_selection}
\title{Model Selection}
\usage{
model_selection(
  loglik,
  n,
  g = 4,
  d = 2,
  modelname = c("IMIX_ind", "IMIX_ind_unrestrict", "IMIX_cor_twostep", "IMIX_cor",
    "IMIX_cor_restrict")
)
}
\arguments{
\item{loglik}{Full log likelihood, result output from IMIX or a sub model in IMIX: `Full MaxLogLik final`}

\item{n}{Total number of genes}

\item{g}{Number of components}

\item{d}{Number of data types}

\item{modelname}{The model name, default is IMIX_ind}
}
\value{
AIC/BIC values of the target model
}
\description{
Model selection based on AIC and BIC information criteria for models in IMIX
}
