% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_component.R
\name{plot_component}
\alias{plot_component}
\title{Plot the Component Selection}
\usage{
plot_component(res_select, type = c("AIC", "BIC"))
}
\arguments{
\item{res_select}{Result output from function model_selection_component()}

\item{type}{Which information criteria to use for plot}
}
\value{
Plot for the model selection of components
}
\description{
Plot the Model selection for components based on AIC and BIC information criteria for models in IMIX
}
