% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FDR_control_adaptive.R
\name{FDR_control_adaptive}
\alias{FDR_control_adaptive}
\title{The adaptive procedure for global FDR control}
\usage{
FDR_control_adaptive(lfdr, alpha)
}
\arguments{
\item{lfdr}{Local FDR for each gene of the mixture model results for one component}

\item{alpha}{Prespecified FDR control level}
}
\value{
The estimated mFDR for the target component and whether the genes is classified in this component after FDR control at alpha level, 1 is yes, 0 is no.
}
\description{
The adaptive procedure for global FDR control based on the output from IMIX models, this can be directly performed by IMIX function, however, if you are interested in other mixture models, this function would be useful.
}
